fix: resolve fast refresh warnings by separating non-component exports
- Moved button, badge, and toggle variants to separate files
- Separated JoinDialogContext from JoinDialogProvider component
- Moved OnboardingContext to separate file
- Fixed useCallback for formatAmount in NutzapCard
- Updated all imports to use new file structure

Results:
- Fast refresh warnings reduced from 10 to 3
- Total ESLint issues reduced from 12 to 5
- All critical errors remain fixed
- Project builds successfully
