+++ src/App.tsx
@@ -9,7 +9,7 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { NostrLoginProvider } from '@nostrify/react/login';
 import AppRouter from './AppRouter';
 import { useSystemTheme } from '@/hooks/useSystemTheme';
-import { JoinDialogProvider } from '@/components/groups/JoinDialogContext';
+import { JoinDialogProvider } from '@/components/groups/JoinDialogProvider';
 
 // DO NOT MODIFY THIS LIST UNLESS YOU ARE ABSOLUTELY CERTAIN EACH RELAY URL YOU ARE ADDING IS VALID AND THE RELAY IS CURRENTLY ONLINE AND CONFIRMED TO BE FULLY FUNCTIONAL AND WORKING.
 const defaultRelays = [
